% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingframes-data.R
\docType{data}
\name{samplingframes}
\alias{samplingframes}
\title{Data set on human inductive reasoning}
\format{a tibble}
\source{
https://psyarxiv.com/2m83v/
}
\usage{
data(samplingframes)
}
\description{
Experiment 2 from Hayes et al (in press).
}
\references{
Hayes et al
}
\keyword{datasets}
