% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{show_environment}
\alias{show_environment}
\title{Displays the contents of the calling environment}
\usage{
show_environment()
}
\value{
Something
}
\description{
Displays the contents of the calling environment
}
\examples{
seeker <- 1
lover <- "keeper"
show_environment()
}
