% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{show_workspace}
\alias{show_workspace}
\title{Displays the contents of the workspace}
\usage{
show_workspace()
}
\value{
Something
}
\description{
Displays the contents of the workspace
}
\examples{
workspace()
}
