% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_independent}
\alias{ttest_independent}
\title{Independent samples t-test}
\usage{
ttest_independent(data, outcome, group)
}
\arguments{
\item{data}{a data frame or tibble}

\item{outcome}{the outcome variable (quoted)}

\item{group}{the grouping variable (quote)}
}
\description{
Independent samples t-test
}
