% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_onesample}
\alias{ttest_onesample}
\title{One sample t-test}
\usage{
ttest_onesample(data, outcome = NULL, null_mean = NULL,
  alternative = "two.sided", ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{outcome}{the outcome variable (quoted)}

\item{null_mean}{the fixed mean to test against (numeric)}

\item{alternative}{character specifying "two.sided" (the default), "greater", or "less"}

\item{...}{other arguments to be passed to t.test}
}
\description{
One sample t-test
}
