% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_paired}
\alias{ttest_paired}
\title{Paired samples t-test}
\usage{
ttest_paired(data, formula = NULL, outcome = NULL, group = NULL,
  id = NULL, alternative = NULL, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{formula}{the model formula (i.e., outcome ~ group)}

\item{outcome}{the outcome variable (quoted)}

\item{group}{the grouping variable (quoted)}

\item{id}{the id variable (qouoted)}

\item{alternative}{an expression specifying the null hypothesis (quoted) or FALSE (default, indicates two sided)}

\item{...}{other arguments to be passed to t.test}
}
\description{
Paired samples t-test
}
