% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_twosample}
\alias{ttest_twosample}
\title{Two sample t-test}
\usage{
ttest_twosample(data, formula = NULL, outcome = NULL, group = NULL,
  test_greater = NULL, equal_variances = FALSE, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{formula}{the model formula (i.e., outcome ~ group)}

\item{outcome}{the outcome variable (quoted)}

\item{group}{the grouping variable (quoted)}

\item{test_greater}{group proposed to be larger under the alternative (character), or NULL (the default, for a two-sided test)}

\item{equal_variances}{should the test assume equality of variance? (default = FALSE)}

\item{...}{other arguments to be passed to t.test}
}
\description{
Two sample t-test
}
