% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace}
\alias{workspace}
\title{Prints the contents of the global environment}
\usage{
workspace()
}
\value{
Something
}
\description{
Prints the contents of the global environment
}
\examples{
workspace()
}
