% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_twosample}
\alias{ttest_twosample}
\title{Two samples t-test}
\usage{
ttest_twosample(data, formula = NULL, outcome = NULL, group = NULL)
}
\arguments{
\item{data}{a data frame or tibble}

\item{formula}{the model formula (i.e., outcome ~ group)}

\item{outcome}{the outcome variable (quoted)}

\item{group}{the grouping variable (quoted)}
}
\description{
Two samples t-test
}
